Add "follow: yes" to the template task in the mssql and elasticsearch subrole.
bz#2058655

The config files are symlinks targeting to the ones in /etc/pcp/subrole.
E.g.,
lrwxrwxrwx. ... /var/lib/pcp/pmdas/mssql/mssql.conf -> ../../../../../etc/pcp/mssql/mssql.conf

The "dest" path in the template module stores the symlink path, and the
symlink is not followed in the template module by default. That makes
the symlink a regular file, while it leaves the target intact.