feat: support for ostree systems
Feature: Allow running and testing the role with ostree managed nodes.

Reason: We have users who want to use the role to manage ostree
systems.

Result: Users can use the role to manage ostree managed nodes.

Add dependency on `ansible.posix` which provides the `rhel_rpm_ostree`
package manager.

When using the `package:` module, add `use:` to force the use of
the ostree package manager on ostree systems, and `omit` if not.

Signed-off-by: Rich Megginson <rmeggins@redhat.com>