fix: pmie webhook configuration for unconfigured system
If pmie never ran before, then /var/lib/pcp/config/pmie/config.default
does not exist, and the `lineinfile` check mode gives `item.found == -1`
instead of `0`. In that case we still want to configure the webhook.

This was previously hidden as the tests run on an unclean VM (from
previous tests) which caused the config file to be written. But it fails
when either running linux-system-roles/metrics'
tests_verify_notification.yml in isolation, or running against a
container build target where services don't run.